Update AIModule.cpp
Changed numberOfShots-multiplier for available TU/TU-cost back to int since 1.9 still is 1.0, when it comes to how often you can attack. And that's obviously way worse.

+++ src/Battlescape/AIModule.cpp
@@ -4553,7 +4553,7 @@ float AIModule::brutalScoreFiringMode(BattleAction *action, BattleUnit *target, 
 		return 0;
 	}
 
-	numberOfShots *= (float)tuTotal / tuCost;
+	numberOfShots *= tuTotal / tuCost;
 	if (numberOfShots < 1)
 		return 0;
